requirements.txt
langgraph==0.2.68
langchain==0.3.15
langchain-openai==0.3.2
langsmith==0.3.2
python-dotenv==1.0.1
requests==2.32.3
streamlit==1.41.1
langchain-google-community[gmail]==2.0.4
google-api-python-client==2.160.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
pandas==2.2.3
unstructured==0.16.22
unstructured[md]==0.16.22
libmagic==1.0
langchain-chroma==0.2.2
IPython==8.30.0
streamlit-mic-recorder==0.0.8
google-cloud-speech==2.31.0
google-cloud-texttospeech==2.25.0
pydub==0.25.1
streamlit-modal==0.1.2
azure-cognitiveservices-speech==1.42.0